+++ packages/unlock-js/src/PublicLock/v10/getPurchaseKeysArguments.js
@@ -1,4 +1,4 @@
-import { ZERO } from '../../constants'
+import { MAX_UINT, ZERO } from '../../constants'
 import utils from '../../utils'
 import formatKeyPrice from '../utils/formatKeyPrice'
 
@@ -28,19 +28,36 @@ export default async function getPurchaseKeysArguments({
   // we parse by default a length corresponding to the owners length
   const defaultArray = Array(owners.length).fill(null)
 
+  const data = (_data || defaultArray).map((d) => d || '0x')
+
   const keyPrices = await Promise.all(
-    (_keyPrices || defaultArray).map(async (kp) => {
-      if (!kp) {
-        // We might not have the keyPrice, in which case, we need to retrieve from the the lock!
-        // TODO: consider using `keyPurchasePriceFor`
-        return await lockContract.keyPrice()
+    Array.from({ length: owners.length }).map(async (_, index) => {
+      const keyPrice = _keyPrices && _keyPrices[index]
+      const owner = owners[index]
+      let referrer = _referrers && _referrers[index]
+      let dataValue = data[index]
+
+      if (!referrer) {
+        referrer = ZERO
       }
-      return formatKeyPrice(kp, erc20Address, decimals, this.provider)
+
+      if (!dataValue) {
+        dataValue = '0x'
+      }
+
+      if (!keyPrice) {
+        // We might not have the keyPrice, in which case, we need to retrieve from the lock!
+        if (owner) {
+          return await lockContract.purchasePriceFor(owner, referrer, dataValue)
+        } else {
+          return await lockContract.keyPrice()
+        }
+      }
+      return formatKeyPrice(keyPrice, erc20Address, decimals, this.provider)
     })
   )
   const keyManagers = (_keyManagers || defaultArray).map((km) => km || ZERO)
   const referrers = (_referrers || defaultArray).map((km) => km || ZERO)
-  const data = (_data || defaultArray).map((d) => d || '0x')
 
   if (
     !(
@@ -60,12 +77,26 @@ export default async function getPurchaseKeysArguments({
   let totalAmountToApprove = totalApproval
 
   if (!totalAmountToApprove) {
-    // total amount to approve
-    totalAmountToApprove = recurringPayments
-      ? keyPrices // for reccuring payments
-          .map((kp, i) => kp * recurringPayments.map(BigInt)[i])
-          .reduce((total, approval) => total + approval, BigInt(0))
-      : totalPrice
+    totalAmountToApprove = keyPrices
+      .map((keyPrice, i) => {
+        if (keyPrice > 0) {
+          const recurringPayment = recurringPayments && recurringPayments[i]
+          if (!recurringPayment) {
+            return keyPrice
+          } else if (recurringPayment === Infinity) {
+            return MAX_UINT
+          } else {
+            return keyPrice * BigInt(recurringPayments)
+          }
+        }
+        return BigInt(0)
+      })
+      .reduce((total, approval) => {
+        if (total === MAX_UINT || approval === MAX_UINT) {
+          return MAX_UINT
+        }
+        return total + approval
+      }, BigInt(0))
   }
 
   return {