+++ unlock-app/src/hooks/useEmailSubscription.ts
@@ -7,23 +7,23 @@ interface Options {
 }
 
 export const useEmailListUnSubscribe = ({ lockAddress, network }: Options) => {
-  const mutation = useMutation(
-    ['unsubscribe', lockAddress, network],
-    async () => {
+  const mutation = useMutation({
+    mutationKey: ['unsubscribe', lockAddress, network],
+    mutationFn: async () => {
       const response = await locksmith.unsubscribeEmail(network, lockAddress)
       return response.data.success
-    }
-  )
+    },
+  })
   return mutation
 }
 
 export const useEmailListReSubscribe = ({ lockAddress, network }: Options) => {
-  const mutation = useMutation(
-    ['resubscribe', lockAddress, network],
-    async () => {
+  const mutation = useMutation({
+    mutationKey: ['resubscribe', lockAddress, network],
+    mutationFn: async () => {
       const response = await locksmith.reSubscribeEmail(network, lockAddress)
       return response.data.success
-    }
-  )
+    },
+  })
   return mutation
 }